requirements.txt
pytest
allure-pytest
selene>=2.0.b4
allure-python-commons~=2.13.5